frontend/backend/requirements.txt
Flask==2.0.2
Werkzeug==2.0.2
SQLAlchemy==1.3.8
Flask-SQLAlchemy==2.4.0
numpy==1.24.0
cryptography==38.0.4
psycopg2-binary==2.9.5
wheel==0.40.0
selenium==4.9.1
mysqlclient==2.1.0
PyMySQL==1.0.2
openpyxl==3.1.1
pandas==1.5.3